+++ packages/core-utils/src/math.test.ts
@@ -3,7 +3,6 @@ import { describe, expect, test } from 'vitest';
 
 import {
   average,
-  fractionDecimalHint,
   fractionIsInteger,
   fractionToDecimal,
   gcd,
@@ -79,23 +78,6 @@ describe('fractionIsInteger', () => {
 
   test('edge case: 0', () => {
     expect(fractionIsInteger(new Fraction(0))).toBe(true);
-  });
-});
-
-describe('fractionDecimalHint', () => {
-  test('returns undefined for fractions that reduce to an integer', () => {
-    expect(fractionDecimalHint(new Fraction(3))).toBeUndefined();
-    expect(fractionDecimalHint(new Fraction(4, 2))).toBeUndefined();
-    expect(fractionDecimalHint(new Fraction(0))).toBeUndefined();
-  });
-
-  test('returns the decimal for fractions with a fractional part', () => {
-    expect(fractionDecimalHint(new Fraction(5, 2))).toBe('2.5');
-    expect(fractionDecimalHint(new Fraction(1, 3))).toBe('~0.333');
-  });
-
-  test('respects the requested precision', () => {
-    expect(fractionDecimalHint(new Fraction(1, 3), 1)).toBe('~0.3');
   });
 });
