+++ packages/magic-products/src/min-spanning-trees/MSTCost.vue
@@ -1,8 +1,11 @@
 <script setup lang="ts">
   import { nullThrows } from '@core/utils/assert';
-  import { fractionDecimalHint } from '@core/utils/math';
   import Well from '@magic/shared/Well';
-  import { Explainer, ExplainerText } from '@magic/shared/explainer';
+  import {
+    Explainer,
+    ExplainerText,
+    fractionDecimalHint,
+  } from '@magic/shared/explainer';
   import { useProvidedGraph } from '@magic/shared/graph-product';
   import Fraction from 'fraction.js';
