+++ packages/magic-shared/src/explainer/index.ts
@@ -1,2 +1,3 @@
 export { default as ExplainerText } from './ExplainerText.vue';
+export { fractionDecimalHint } from './fractionDecimalHint.ts';
 export type { Explainer, ExplainerHighlight } from './types.ts';
